// packages/ui-kit/src/utils/popper/types.ts
export type Position = 'above' | 'below' | 'before' | 'after';
export type Alignment = 'top' | 'bottom' | 'start' | 'end' | 'center';

export type AutoSize = 'height' | 'height-always' | 'width' | 'width-always' | 'always' | boolean;

export type PopperChildrenFn = (props: PopperChildrenProps) => React.ReactElement;

export type Boundary = PopperBoundary | 'scrollParent' | 'window';